Delete obsolete test files for descriptor mappings
The removed files contained tests for descriptor mappings that are no longer relevant or used. This cleanup reduces maintenance overhead and improves the clarity of the codebase. All tests targeted specific mapping implementations, which are now unnecessary.

+++ wgpu4k/src/jsMain/kotlin/io.ygdrasil.wgpu/mapper/SamplerDescriptor.kt
@@ -15,5 +15,5 @@ internal fun map(input: SamplerDescriptor): GPUSamplerDescriptor = createJsObjec
     lodMinClamp = input.lodMinClamp
     lodMaxClamp = input.lodMaxClamp
     if (input.compare != null) compare = input.compare.stringValue
-    maxAnisotropy = input.maxAnisotropy
+    maxAnisotropy = input.maxAnisotropy.toShort()
 }
